初步完成好友分组RelationGroupInfo模块 进行了基本测试,api的设计可能需要优化 FriendApply的group也更新为Long类型 通用异常处理模块的未知异常更改为会打印trace 更新ucc模块redis查询在线状态 uit所有基本模块大致完成！

+++ friend-apply/src/main/java/com/syh/uit/friend_apply_service/model/FriendApply.java
@@ -5,7 +5,7 @@ public class FriendApply {
     private final int applyID;
     private final int targetID;
     private String petName;
-    private int group;
+    private long group;
     private String comment;
     private final long applyTime;
     private int state;
@@ -29,7 +29,7 @@ public class FriendApply {
         }
     }
 
-    public FriendApply(long applicationID, int applyID, int targetID, String petName, int group, String comment, long applyTime, int state) {
+    public FriendApply(long applicationID, int applyID, int targetID, String petName, long group, String comment, long applyTime, int state) {
         this.applicationID = applicationID;
         this.applyID = applyID;
         this.targetID = targetID;
@@ -57,7 +57,7 @@ public class FriendApply {
         return petName;
     }
 
-    public int getGroup() {
+    public long getGroup() {
         return group;
     }
